rename /common/errorFilter folder to /common/exceptionFilter
